fix: attempt to fix 404 error on providers page
This fix isn't definitive. The problem is that when fetching the token when loading the providers page, sometimes the token doesn't exist, and sometimes it's simply not found in the blockchain at the moment of fetching. The second *SHOULD* never happen, since we only navigate naturally to the providers page after the transactions have been completed. This leads me to believe the issue is with the metamask provider and its ridiculous caching strategies. The proposed fix uses the web3 websocket provider to fetch the node details, but it's not 100% reliable. The definitive fix would be to use the upcoming middleware to store registered tickers, so then if the fetch fails for a ticker we know has been registered, we can report this error as a problem with the blockchain, or simply retry the fetch a couple of times until it works and THEN report the error. Right now we have no surefire way of differentiating between a nonexistent ticker and a failed fetch of an existing one.