fix(ui): set account and network polling interval to 1 second
Metamask suggested polling the node every 100ms to check if the user had changed networks or accounts in order to reload the dApp. This, of course, is utterly insane. Some versions of chrome are sensible enough to throttle this insanity, but firefox received the full brunt of this onslaught of requests and produced weird behavior (even causing errors in ganache-cli). I suspect this might have even been causing some other issues. I changed the polling interval to 1 second (which may be slower to detect a network or account change but at least it won't crash the user's browser. This at least fixes weird behavior in firefox and possibly other performance/network issues.